requirements.txt
python-binance
requests
incremental
